+++ web/src/components/ui/index.ts
@@ -3,7 +3,7 @@ export { default as Avatar } from './Avatar';
 export { default as BranchBadge } from './BranchBadge';
 export { default as Button } from './Button';
 export { default as Card } from './Card';
-export { default as Badge, LabelBadge, LabelColorsContext, HideDefaultLabelsContext, DefaultLabelsContext, StatusBadge } from './Badge';
+export { default as Badge, LabelBadge, LabelIndicator, LabelColorsContext, HideDefaultLabelsContext, DefaultLabelsContext, StatusBadge } from './Badge';
 export { default as Modal, ModalFooter } from './Modal';
 export { default as Progress, ProgressRing } from './Progress';
 export { default as StatusIcon } from './StatusIcon';